Separate logic of prometheus response parsing with its service.
